- project 중복 체크 버그 수정 - runtime 유효성 체크 추가

+++ app/flask_api/runtime_helper.py
@@ -12,6 +12,7 @@ def getRuntimePathAndImage(runtime, model):
     if rows is not None:
         if len(rows) != 0:
             return rows[0]['path'], rows[0]['image_name'], rows[0]['cuda_path'], rows[0]['cudnn_path']
+    return None, None, None, None
 
 
 def getTensorRTPath(runtime, tensorRT):
@@ -27,6 +28,9 @@ def getTensorRTPath(runtime, tensorRT):
 def getBasicYaml(userLoginID, userName, projectName, projectID, nodeID, runtime, model, tensorRT, framework, inputPath, outputPath):
     runtimePath, imageName, cudaPath, cudnnPath = getRuntimePathAndImage(runtime, model)
     tensorRTPath = getTensorRTPath(runtime, tensorRT)
+
+    if runtimePath is None or imageName is None or cudnnPath is None or cudaPath is None or tensorRTPath is None:
+        return None
 
     data = {'apiVersion': 'v1', 'kind': 'Pod',
             'metadata': {'name': nodeID, 'labels': {'app': 'nfs-test'}},
@@ -128,6 +132,8 @@ def getBasicPVCYaml(userLoginID, projectName):
 
 def makeYamlTrainRuntime(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, outputPath):
     data = getBasicYaml(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, outputPath)
+    if data is None:
+        return None
     data['spec']['containers'][0]['args'][0] += 'rm -rf ' + pathJoin('/root/user/', outputPath) \
                                                 + '; nohup python train.py --project /root/user --name ' + outputPath \
                                                 + ' --data  ' + pathJoin(datasetPath + '/dataset.yaml') \
@@ -136,6 +142,8 @@ def makeYamlTrainRuntime(userLoginID, userName, projectName, projectID, node_id,
 
 def makeYamlValidateRuntime(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, modelPath, outputPath):
     data = getBasicYaml(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, outputPath)
+    if data is None:
+        return None
     data['spec']['containers'][0]['args'][0] += 'rm -rf ' + pathJoin('/root/user/', outputPath) \
                                                 + '; nohup python val.py --project /root/user --name ' + outputPath \
                                                 +  ' --data ' + pathJoin(datasetPath + '/dataset.yaml') + ' --device 0 --weights ' + modelPath \
@@ -144,6 +152,8 @@ def makeYamlValidateRuntime(userLoginID, userName, projectName, projectID, node_
     return data
 def makeYamlOptimizationRuntime(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, modelPath):
     data = getBasicYaml(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, modelPath, modelPath)
+    if data is None:
+        return None
     data['spec']['containers'][0]['args'][0] += 'nohup python export.py --weights ' + modelPath \
                                                 + ' --include engine --device 0 --half --batch-size 1 --imgsz 640 --verbose &>> /root/user/logs/' + node_id + '.log'
 
@@ -151,6 +161,8 @@ def makeYamlOptimizationRuntime(userLoginID, userName, projectName, projectID, n
 
 def makeYamlOptValidateRuntime(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, modelPath, outputPath):
     data = getBasicYaml(userLoginID, userName, projectName, projectID, node_id, runtime, model, tensorRT, framework, datasetPath, outputPath)
+    if data is None:
+        return None
     data['spec']['containers'][0]['args'][0] += 'rm -rf ' + pathJoin('/root/user/', outputPath) \
                                                 + '; nohup python val.py --project /root/user --name ' + outputPath \
                                                 +  ' --weights ' + modelPath + ' --data ' + pathJoin(datasetPath, '/dataset.yaml') \
